Removed cruft
