+++ src/view/camera.ts
